Create log-reg-measurements.py
Please excuse the paths. I have yet to figure out how to make paths more generic. You can, however, just copy in your own path to the data and it will work, just not pretty or efficient. Here is the code I used to get the ORs we presented today. As I said during our last meeting, I simply changed the variable names for lactate bins and used the length of the races list instead, so I didn't have to figure out the logic of the loops. I probably need a couple more hours to figure that out myself. As Leo pointed out, there might be some issues with the fundamentals of how this is produced...